Made different component for receive using qr code
