+++ src/components/WalletPage/WalletDetail.tsx
@@ -1,47 +1,26 @@
 'use client'
-import {
-  ArrowDownUp,
-  Check,
-  Copy,
-  CreditCard,
-  Download,
-  Info,
-  Send,
-  X,
-} from 'lucide-react'
-import { QRCodeSVG } from 'qrcode.react'
 import React, { useState } from 'react'
-import { actions } from './actions'
+import { actions, ActionType } from './actions'
+import { ReceiveQR } from './ReceiveQR'
 
 
-interface WalletDetailProps {
+export interface WalletDetailProps {
   wallet?: string
   balance?: number
 }
 
 const WalletDetail = ({ wallet, balance }: WalletDetailProps) => {
-  const [showQR, setShowQR] = useState(false)
-  const [copied, setCopied] = useState(false)
+  const [currentAction, setCurrentAction] = useState<null | ActionType>(null);
 
-  const handleCopy = () => {
-    navigator.clipboard.writeText(wallet ?? 'no wallet address')
-    setCopied(true)
-    setTimeout(() => setCopied(false), 2000)
-  }
-  const handleReceive = () => {
-    setShowQR(true)
-  }
+  const handleActionClick = (action: ActionType) => {
+   setCurrentAction(action)
+  };
+
 
   const handleClose = () => {
-    setShowQR(false)
+    setCurrentAction(null);
   }
 
-  const formatWallet = () => {
-    if (wallet) {
-      return `${wallet?.slice(0, 4)}...${wallet?.slice(-4)}`
-    }
-    return 'no wallet address'
-  }
 
   return (
     <div className="rounded-[22px] flex flex-col items-center gap-3 sm:w-[450px] border-2 p-7 sm:p-10 shadow-md">
@@ -68,7 +47,7 @@ const WalletDetail = ({ wallet, balance }: WalletDetailProps) => {
             <div key={index} className="flex flex-col items-center gap-1">
               <button
                 className="flex justify-center items-center w-[66px] h-[67px] bg-black opacity-80 rounded-[30%] text-white"
-                onClick={action.label === 'Receive' ? handleReceive : undefined}
+                onClick={()=> handleActionClick(action.type)}
               >
                 {<action.icon/>}
               </button>
@@ -103,52 +82,7 @@ const WalletDetail = ({ wallet, balance }: WalletDetailProps) => {
           </button>
         </div>
       </div>
-      {showQR && (
-        <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50">
-          <div className="bg-white rounded-lg p-6 shadow-lg max-w-sm w-full">
-            <div className="flex justify-between items-center">
-              <h3 className="text-lg font-bold">Your Wallet Address</h3>
-              <button
-                onClick={handleClose}
-                className="text-gray-500 hover:text-gray-800"
-              >
-                <X size={25} />
-              </button>
-            </div>
-            <div className="flex flex-col items-center gap-4 mt-4">
-              <QRCodeSVG
-                value={wallet ?? 'no wallet address found'}
-                size={180}
-              />
-            </div>
-            <div
-              className="flex items-center justify-evenly bg-gray-100 shadow-md shadow-gray-200 p-2 rounded-md w-full mt-4"
-              onClick={handleCopy}
-            >
-              <p className="font-mono text-gray-700 break-words">
-                {formatWallet()}
-              </p>
-              <div className="relative">
-                <button
-                  className={`bg-blue-500 text-white p-2 rounded-full hover:bg-blue-600 cursor-pointer transition duration-300`}
-                >
-                  {copied ? <Check size={16} /> : <Copy size={16} />}
-                </button>
-                {copied && (
-                  <div className="absolute top-full left-0 mt-1 bg-gray-800 text-white text-xs py-1 px-2 rounded-md">
-                    Copied!
-                  </div>
-                )}
-              </div>
-            </div>
-
-            <p className="text-gray-500 text-xs mt-2 flex gap-1 justify-center w-full whitespace-nowrap">
-              <Info/>
-              <i className='mt-1'>Only send crypto to this address via the Solana network.</i>
-            </p>
-          </div>
-        </div>
-      )}
+      {currentAction === 'receive' && <ReceiveQR wallet={wallet} onClose={handleClose}/>}
     </div>
   )
 }
